requirements.txt
django==5.0
djangorestframework==3.14.0
environs[django]==9.5.0